requirements.txt
pandas>=1.5.0
mysql-connector-python>=8.0.0
requests>=2.28.0
openpyxl>=3.0.0
xlrd>=2.0.0
python-dotenv>=0.21.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
